requirements.txt
pandas
numpy
requests
python-dotenv
streamlit
scikit-learn
ta
alpaca-trade-api
pytest
